renamed packages & WIP fixes
