+++ src/main/java/dao/CourseDAO.java
@@ -1,14 +1,14 @@
 package dao;
 
-import models.Course;
-import models.Customer;
+import domainModel.Course;
+import domainModel.Customer;
 
 import java.util.List;
 
 public interface CourseDAO extends DAO<Course, Integer> {
     public int getNextID();
 
-    public List<Customer> getCustomersOfCourse(Integer courseId);
+    public List<Customer> getCustomersOfCourse(Integer courseId); // getAttendees
 
     public List<Course> getCoursesOfCustomer(String fiscalCode);
