+++ src/main/java/domainModel/Person.java
@@ -1,4 +1,4 @@
-package models;
+package domainModel;
 
 import java.util.Objects;
