+++ src/main/java/domainModel/membership/Membership.java
@@ -1,4 +1,4 @@
-package models.membership;
+package domainModel.membership;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
