+++ src/test/java/dao/SQLiteTrainerDAOTest.java
@@ -1,6 +1,6 @@
 package dao;
 
-import models.Trainer;
+import domainModel.Trainer;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
